The first version of ikileme database
The database contains 420 findings, a couple of them are incorrect.
The script goes through a database of all words and checks if a finding in a database contains the first word more than once. It is a problem for findings like: "hem ... hem ..." or "bir boydan bir boya", There are less than 10 false positives